Move helper and customview to util package
