+++ app/src/main/kotlin/com/ayatk/biblio/ui/fragment/NovelInfoFragment.kt
@@ -12,7 +12,7 @@ import android.view.ViewGroup
 import com.ayatk.biblio.databinding.FragmentNovelInfoBinding
 import com.ayatk.biblio.model.Novel
 import com.ayatk.biblio.repository.library.LibraryRepository
-import com.ayatk.biblio.ui.helper.Navigator
+import com.ayatk.biblio.ui.util.helper.Navigator
 import com.ayatk.biblio.viewmodel.NovelInfoViewModel
 import org.parceler.Parcels
 import javax.inject.Inject
