+++ app/src/main/kotlin/com/ayatk/biblio/ui/util/helper/BottomNavigationViewHelper.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2016-2017 Aya Tokikaze. All Rights Reserved.
  */
 
-package com.ayatk.biblio.ui.helper
+package com.ayatk.biblio.ui.util.helper
 
 import android.support.design.internal.BottomNavigationItemView
 import android.support.design.internal.BottomNavigationMenuView
